Created a RegisterStep4 component and added it to the flow of the parent registration flow, reordered the parent registration flow so the number of students is set after the parent fills out the parent info. Also changed all the refrences to the old components to the new components
