+++ src/components/pages/Registration/RegisterStep1.js
@@ -2,7 +2,7 @@ import React from 'react';
 import RegistrationProgress from './RegistrationProgress';
 import '../../../styles/registration.less';
 
-export default function ParentWelcome() {
+export default function RegisterStep1() {
   return (
     <div className="reg-content-container">
       <RegistrationProgress step_num={0} />
@@ -12,10 +12,6 @@ export default function ParentWelcome() {
           You are steps away from unleashing your superpowers
         </p>
         <a href="/register-2">CREATE PROFILES</a>
-      </div>
-      <div className="footer-content">
-        <p>Looking for Instructor Registration?</p>
-        <a href="/instructor-register-1">Click Here</a>
       </div>
     </div>
   );
