+++ src/styles/ParentStyles/index.less
@@ -65,7 +65,7 @@
 }
 
 .dropStudent {
-  background-color: @parent-navbar__bg--white;
+  background-color: white;
   width: 25%;
   display: flex;
   align-items: center;
@@ -117,16 +117,7 @@
 }
 }
 
-.tab_container {
-  display: flex;  
-  background-color: #edeed7;  
-}
-.tab_container > div {
-  background-color: #f1f1f1;
-  margin: 20px 0px  20px 50px;
-  
-  font-size: 30px;
-}
+
 
 .card-name {
   font-size: 1rem;
@@ -177,13 +168,12 @@
 }
 
 .card {
-
-}
-.tab_container{
   display: flex;
-  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  width: 100%;
 }
-
 
 .add-student-button {
   background-color: @btn__bg--orange;
@@ -241,7 +231,7 @@
 
 .dropStudent {
   padding: 12% 30%;
-  background-color: @parent-navbar__bg--white;
+  background-color: white;
 }
 
 //****Parent-Tasks styles****
@@ -281,31 +271,6 @@
   justify-content: center;
   margin-top: 5%;
   border-radius: 5%;
-}
-
-//****Parent-Resources styles****
-
-.cardContainer {
-  display: flex;
-  flex-direction: row;
-  flex-flow: nowrap;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.artImg {
-  width: 100%;
-  padding-bottom: 5%;
-}
-
-.resourceCard {
-  width: 25%;
-  height: 25vh;
-  display: flex;
-  flex-direction: column;
-  flex-flow: nowrap;
-  justify-content: center;
-  margin-top: 5%;
 }
 
 // Create New Student Modal
@@ -354,6 +319,11 @@
   display: flex;
   flex-direction: column;
   margin: 1em;
+}
+
+.create-new-student-notes {
+  width: 100%;
+  margin-top: 1rem;
 }
 
 .create-new-student-submit {
